+++ tests/TestCase/GlobalFunctionsTest.php
@@ -69,6 +69,40 @@ class GlobalFunctionsTest extends TestCase
 
     /**
      * @test
+     * @uses \is_url()
+     */
+    public function testIsUrl(): void
+    {
+        foreach ([
+                     'https://www.example.com',
+                     'http://www.example.com',
+                     'www.example.com',
+                     'http://example.com',
+                     'http://example.com/file',
+                     'http://example.com/file.html',
+                     'www.example.com/file.html',
+                     'http://example.com/subdir/file',
+                     'ftp://www.example.com',
+                     'ftp://example.com',
+                     'ftp://example.com/file.html',
+                     'http://example.com/name-with-brackets(3).jpg',
+                 ] as $url) {
+            $this->assertTrue(is_url($url), 'Failed asserting that `' . $url . '` is a valid url');
+        }
+
+        foreach ([
+                     'example.com',
+                     'folder',
+                     DS . 'folder',
+                     DS . 'folder' . DS,
+                     DS . 'folder' . DS . 'file.txt',
+                 ] as $url) {
+            $this->assertFalse(is_url($url));
+        }
+    }
+
+    /**
+     * @test
      * @uses rtr()
      */
     public function testRtr(): void